feat: album 관련 API에서 에러 코드 관련 중복되는 부분 모듈화
